Rest query for all reservations and reservation by ID currently working Need to still develop rest for new reservation

+++ src/main/java/carRental/domain/CarReservationRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
 public interface CarReservationRepository extends CrudRepository<CarReservation, Long> {
-//  @Query()
-  CarReservation findByCustomerID(int resID);
+
+  CarReservation findByCustomerID(long resID);
 
   Iterable<CarReservation> findAll();
